refactor(assembly): flatten crate structure, export linker components
