+++ assembly/src/instruction/crypto_ops.rs
