+++ miden-vm/tests/integration/cli/data/kernel_main.masm
@@ -1,7 +1,7 @@
 proc.internal_proc
     caller
     drop
-    exec.::kernel::lib::lib_proc
+    exec.::$kernel::lib::lib_proc
 end
 
 export.kernel_proc
